pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "writing-tools-mcp"
version = "0.1.0"
description = "MCP server with a PySide GUI for text analysis tools."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
   { name="Will McGinnis", email="will@mcginniscommawill.com" },
]

dependencies = [
    "mcp[cli]",
    "pyspellchecker",
    "textstat",
    "spacy",
    "markdown-it-py",
    "Pyside6",
    "requests",
    "transformers>=4.35.0",
    "torch>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0"
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[project.urls]
"Homepage" = "https://github.com/wdm0006/writing-tools-mcp" 
"Bug Tracker" = "https://github.com/wdm0006/writing-tools-mcp/issues" 

# Example for ruff (linter) configuration
[tool.ruff]
line-length = 120
select = ["E", "W", "F", "I", "C", "B"] 
ignore = ["E501", "C901"]

[tool.ruff.format]
quote-style = "double"

[tool.hatch.build.targets.wheel]
packages = ["app", "server"]